refactor(WinningResult): setPrizeMoney() 메서드의 파라미터 정수형 타입 변경
- 기존 int 타입으로 받아오던 prizeMoney를 테스트를 위해 long 타입으로 변경

+++ src/main/java/lotto/domain/WinningResult.java
@@ -85,7 +85,7 @@ public class WinningResult {
         return (double) prizeMoney / amount * 100;
     }
 
-    public void setPrizeMoney(int prizeMoney) {
+    public void setPrizeMoney(long prizeMoney) {
         this.prizeMoney = prizeMoney;
     }
 }
